drmu: Change find plane behaviour to take a ref
The only reason to find a new plane is if you want to use it so claim as
part of the search. This is more efficient than fidinging then claiming
then retrying if the claim fails. This also means that the output plane
claims won't fail when racing against another thread doing the same
which they could previously.