hello_drmu: Rework to allow playing of multiple videos
Rework to separate out player code s.t. it has a proper environment and
having done that put it in its own source file.
Rework arument parsing to allow for multiple windows with their own
playlists. This is still imperfect as many options are still per
playlist and really want to be per flle.