drmu_poll: Rework queueing, adds queues, merge modes, uses blocking commits
Substantial rework of the way queueing works internally. External
API should be pretty the same.

Biggest structural change is to use a blocking commit (per Q) in
its own thread rather than waiting for flip events. This was done
due to flip events sometimes happening before the commit has actually
finished - mainly when a mode change is part of the atomic which
can happen frequently on writeback.

Multiple Qs have been added to make it easier to interleave writebacks
with display, and it will be easier for multiple displays too

Merge modes added to control how multiple pending requests are dealt
with. Writeback most commonly want a replace op, display a merge.

Q locking added to freeze output on writeback Qs until the FD is
signalled. Comments in the driver are clear this is wanted though
with a blocking commit it isn't obvious it is actually needed.